Remove http_client_winhttppal.cpp
